fix CloseIcon & useClasses
